Subset of the dataset aboutthe Deferred Action for Childhood Arrivals (DACA) program

A dataset about Deferred Action for Childhood Arrivals (DACA) program containing 330106 observations

DACAsub

Format

A data frame with 330106 rows and 18 columns:

inschool

High school attendance (In School)

hs

High school graduation (HS Degree)

scol

Some College

post

Post treatment period: a dummy variable that equals 1 on or after 2012

elig

Eligible for treatment: is equal to one for individuals in the sample who arrived in the US by age 10 and by year 2007, and who are currently not citizens.

fem

Female

race

Race

var.bpl

Birth region (Birthplace)

state

State

year

Year

age

Age

yrimmig

Year of immigration

ageimmig

Age of immigration

a1418

Equals 1 if age between 14 and 18

a1922

Equals 1 if age between 19 and 22

a2330

Equals 1 if age between 23 and 30

htus

Available

perwt

Person weight

Source

https://doi.org/10.1257/pol.20180352.