Two Subset of the dataset about the Deferred Action for Childhood Arrivals (DACA) program

Two A dataset about Deferred Action for Childhood Arrivals (DACA) program containing 330106 observations

Usage

Unempl

Format

A data frame with 17379 rows and 11 columns:

FT

2

unemp

2

sex

1 - Male, 2 - Female

statefip

State

school

Post treatment period: a dummy variable that equals 1 on or after 2012

age

Age

yrimmig

Year of immigration

educ

Years of education

period

1-8

treatment_period

Birth region (Birthplace)

eligible

Eligible for treatment