Plotting Heterogenous Treatment Effects

The didnpplothte take output created by didnpreg and produces plots of Heterogeneous Treatment Effects with their corresponding confidence bounds. Average effects per group are produced for discrete covariates and a continuum of effects are produced for continuous covariates.

didnpplothte(
  obj,
  level = 95,
  by,
  n.intervals = 10,
  over = NULL,
  xlab = "",
  ylab = "ATET",
  over.lab = "ATET over",
  over.ci.lab = "Conf.Int.",
  point_size = 3,
  line_width = 2,
  by.labels.values = NULL,
  over.labels.values = NULL,
  text_size = 17
)

Arguments

obj

an object of class "didnp".

by

categorical or continuous.

n.intervals

number of intervals for the numeric "by" to be split into. Default is 10.

over

necessarily categorical.

xlab

Label for horizontal axis. Default is "".

ylab

Label for vertical axis. Default is "ATET".

point_size

of ATET. Default is 3.

line_width

of ATET for the numeric "by". Default is 2.

text_size

for ggplot object. Default is 17.

labels.values

for the categorical "by", the dataframe with two columns. The first column contains unique values of the "by", while the second column contains corresponding values that would appear on a graph. Default is NULL, implying that the unique values of the "by" will be used.

Value

didnpplot returns a dataframe containing containing the following columns:

plot.a	ggplot object

data.a	data used to produce plot.a

plot.b	ggplot object for TTb if TTb = TRUE was used in running didnpreg

data.a	data used to produce plot.b

Details

Vector "by" (and "over") must be of length of TTa.i if TTb = FALSE was used in running didnpreg and of length of TTb.i if TTb = TRUE was used in running didnpreg

References

... (...). This. Journal of , 1(1), 1-1 https://doi.org/10.

Author

Oleg Badunenko oleg.badunenko@brunel.ac.uk,

Daniel J. Henderson djhender@cba.ua.edu,

Stefan Sperlich stefan.sperlich@unige.ch

Examples

if (FALSE) {
  data(DACAsub, package = "didnp")
  # will get a data frame 'DACAsub' with 330106 rows and 18 columns

  # get the subsample
  DACAsub$mysmpl <- mysmpl <-
    DACAsub$a1922==1 & !is.na(DACAsub$a1922) &
    DACAsub$htus==1 & !is.na(DACAsub$htus)

  # generate 'treatment_period'
  DACAsub$treatment_period <- ifelse(DACAsub[,"year"]>2011,1,0)

  # define formula with the weight
  form1 <- inschool ~ fem + race + var.bpl + state + age + yrimmig +
    ageimmig | inschool | year | elig | treatment_period | perwt

  tym1a <- didnpreg(
    form1,
    data = DACAsub,
    subset = mysmpl,
    bwmethod = "opt",
    boot.num = 399,
    TTb = FALSE,
    print.level = 1,
    cores = 4)

  # Print the summary
  summary(tym1a)

}