Treatment Effect Estimators

The didnpreg command contains tools for computing both heterogenous and average treatment effects for the treated in a model-free differences-in-differences framework.

Usage

didnpreg(...)

# S3 method for formula
didnpreg(
  formula,
  data = stop("argument 'data' is missing"),
  subset,
  bwmethod = "opt",
  boot.num = 399,
  TTx = FALSE,
  TTb = TRUE,
  print.level = 1,
  cores = 4,
  seed = 17345168,
  ...
)

# S3 method for default
didnpreg(
  outcome,
  regressors,
  id,
  time,
  treatment,
  treatment_period,
  weights = NULL,
  bwmethod = "opt",
  boot.num = 399,
  TTx = FALSE,
  TTb = TRUE,
  print.level = 1,
  cores = 4,
  seed = 17345168,
  ...
)

Arguments

formula

an object of class formula (or one that can be coerced to that class): a symbolic description of the model. The details of model specification are given under `Details'

data,

# name of the data frame; must be specified if the 'formula' method is used

subset

NULL, optional subsample of 'data'

bwmethod

"opt", # plug-in is rule of thumb for continuous and basic for categorical, can be cross-validation

boot.num

399,

TTx

FALSE

TTb

FALSE

print.level

the level of printing; larger number implies more output is printed. Default is 1. 0 suppresses all printing.

cores

Integer specifies the number of cores to be used for parallel computation.

outcome

a vector, matrix, or data frame of length \(NT\). The outcome can be continuous or dummy.

regressors

a data frame with \(NT\) rows that contains regressors. A data frame is required to identify the type/class of each regressor.

id

a vector, matrix, or data frame of length \(NT\) that identifies the unit of observation.

time

a vector, matrix, or data frame of length \(NT\) that specifies in which period id is observed.

treatment

a vector, matrix, or data frame of length \(NT\) with zeros for the control and ones for the treated observations.

treatment_period

a vector, matrix, or data frame of length \(NT\) with zeros for the period before treatment and ones for the period of treatment and after.

weights

NULL,

Value

didnpreg returns a list containing:

esample	A vector of TRUE/FALSE values identifying observations used in estimation. Relevant for the 'formula' method but complete cases will also be checked in the matrix method

esample_treated	A vector of TRUE/FALSE values identifying observations used in estimation and is treated (intersection of esample and treatment)

esample_TTa	A vector of TRUE/FALSE values identifying observations used in estimation of TTa)

esample_TTb	A vector of TRUE/FALSE values identifying observations used in estimation of TTb. Returned only if TTb = TRUE)

N	number of units of obs, defined by the id

N_dt	the number of observations used as training data for each conditional expectation

T_i	number of observations in each id

bws	bandwidth(s) for the data

TTx	the DiD estimators of the conditional TT (also known as CATET) (equation 2.12)

TTa

TTb	NULL is option TTaOnly is used

Details

All estimations are based. y ~ x1 + x2 | id | time | treatment | treatment_period | weights

References

... (...). This. Journal of , 1(1), 1-1 https://doi.org/10.

Author

Oleg Badunenko oleg.badunenko@brunel.ac.uk, Daniel J. Henderson djhender@cba.ua.edu, Stefan Sperlich stefan.sperlich@unige.ch

Examples

if (FALSE) {
  didnpreg(jhklangkjhakgkjh)
}